+++ modules/bridge/src/main/ts/ephox/bridge/components/menu/FancyMenuItem.ts
@@ -4,6 +4,7 @@ import { Result, Fun } from '@ephox/katamari';
 export interface FancyMenuItemSpec {
   type: 'fancymenuitem';
   fancytype: string;
+  initData?: Record<string, any>;
   onAction: (data: any) => void;
 }
 
@@ -11,6 +12,7 @@ export interface FancyMenuItem {
   type: 'fancymenuitem';
   fancytype: keyof FancyActionArgsMap;
   onAction: <K extends keyof FancyActionArgsMap>(data: FancyActionArgsMap[K]) => void;
+  initData: Record<string, any>;
 }
 
 export interface FancyActionArgsMap {
@@ -23,6 +25,7 @@ const fancyTypes: (keyof FancyActionArgsMap)[] = [ 'inserttable', 'colorswatch' 
 export const fancyMenuItemSchema = StructureSchema.objOf([
   FieldSchema.requiredString('type'),
   FieldSchema.requiredStringEnum('fancytype', fancyTypes),
+  FieldSchema.defaulted('initData', {}),
   FieldSchema.defaultedFunction('onAction', Fun.noop)
 ]);