+++ modules/tinymce/src/plugins/table/test/ts/browser/ui/TableBorderWidthTest.ts
@@ -33,6 +33,7 @@ describe('browser.tinymce.plugins.table.ui.TableBorderWidthTest', () => {
     await pAssertStyleCanBeToggledOnAndOff(hook.editor(), {
       menuTitle: 'Border width',
       subMenuTitle: '1PX',
+      subMenuTitleUndo: 'None',
       checkMarkEntries: 2,
       rows: 1,
       columns: 1,
@@ -44,6 +45,7 @@ describe('browser.tinymce.plugins.table.ui.TableBorderWidthTest', () => {
     await pAssertStyleCanBeToggledOnAndOff(hook.editor(), {
       menuTitle: 'Border width',
       subMenuTitle: '1PX',
+      subMenuTitleUndo: 'None',
       checkMarkEntries: 2,
       rows: 2,
       columns: 2,